Finalize App - test p.1

+++ src/controllers/user.ts
@@ -1,8 +1,12 @@
 import { NextFunction, Request, Response } from 'express';
 import bcrypt from 'bcryptjs';
-import { appErrorHandler, StatusCodeError } from '../helpers/custom-errors';
+import { appErrorHandler, StatusCodeError } from '../handlers/error-handler';
 import User from '../models/user';
 import { applyPatch, Operation } from 'fast-json-patch';
+import {
+  isUpdateOperation,
+  UpdateOperation
+} from '../handlers/patch-user-handler';
 
 export const getUsers = async (
   req: Request,
@@ -11,7 +15,37 @@ export const getUsers = async (
 ) => {
   try {
     const users = await User.getUsers();
-    res.status(200).json({ message: 'Active users list.', users });
+    res.status(200).json({ message: 'Users', users });
+  } catch (err) {
+    const error = appErrorHandler(err);
+    next(error);
+  }
+};
+
+export const getUser = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  const { user } = req;
+  const userId = user.getId();
+
+  try {
+    if (!userId) {
+      throw new StatusCodeError('User Not Found', 404);
+    }
+    const ETag = await user.generateETag();
+    res
+      .status(200)
+      .set('ETag', ETag.toString())
+      .json({
+        user: {
+          id: user.getId(),
+          username: user.getUsername(),
+          created_at: user.getCreatedAt(),
+          active: user.getUserStatus()
+        }
+      });
   } catch (err) {
     const error = appErrorHandler(err);
     next(error);
@@ -44,44 +78,43 @@ export const updateUser = async (
   next: NextFunction
 ) => {
   const { user } = req;
-  const { username, password } = req.body;
-  const patch = [];
+  const patch = req.body;
 
   try {
-    if (typeof username === 'string') {
-      const usernamePatch = {
-        op: 'replace',
-        path: '/username',
-        value: username
-      } as Operation;
-      patch.push(usernamePatch);
-    }
-
-    if (typeof password === 'string') {
-      const isPasswordMatch = await bcrypt.compare(
-        password,
-        user.getPassword()
-      );
-      if (!isPasswordMatch) {
-        const hashedPassword = await bcrypt.hash(password, 12);
-        const passwordPatch = {
-          op: 'replace',
-          path: '/password',
-          value: hashedPassword
-        } as Operation;
-        patch.push(passwordPatch);
-      }
-    }
-
-    if (patch.length === 0) {
-      return res.status(200).json({
-        message: `The entered data is identical to the existing user or data was lack. User remains unchanged.`
-      });
-    }
-    applyPatch(user, patch);
+    const preparedPatchWithHashedPassword = await Promise.all(
+      patch.map(async (operation: UpdateOperation) => {
+        if (operation.path !== '/password') {
+          return operation;
+        }
+        if (operation.path === '/password') {
+          const modifiedOperation = { ...operation };
+          const isPasswordMatch = await bcrypt.compare(
+            operation.value,
+            user.getPassword()
+          );
+          if (isPasswordMatch) {
+            throw new StatusCodeError(
+              'You cannot change the password to the same one you already have.',
+              422
+            );
+          }
+          if (!isPasswordMatch) {
+            const hashedPassword = await bcrypt.hash(operation.value, 12);
+            modifiedOperation.value = hashedPassword;
+            return modifiedOperation;
+          } else {
+            return;
+          }
+        }
+      })
+    );
+    const filteredPatch = preparedPatchWithHashedPassword.filter(
+      (element) => element !== undefined
+    );
+    applyPatch(user, filteredPatch);
     user.save();
     res.status(200).json({
-      message: `User data was changed. Current session for user was closed. Please login again into account.`
+      message: `User data was changed. Current token for user was deleted. Please login again into account.`
     });
   } catch (err) {
     const error = appErrorHandler(err);
@@ -97,6 +130,9 @@ export const deleteUser = async (
   const { user } = req;
 
   try {
+    if (user.isDeleted()) {
+      throw new StatusCodeError('User is deleted already.', 409);
+    }
     const updatedUser = await user.delete();
     const updatedUsername = updatedUser.getUsername();
     if (!updatedUser.isDeleted()) {
